refactor: 레포지토리, 서비스 인터페이스로 추상화
