+++ src/main/java/com/sb02/todoapp/repository/TodoRepository.java
@@ -1,25 +1,10 @@
 package com.sb02.todoapp.repository;
 
 import com.sb02.todoapp.domain.Todo;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public class TodoRepository {
-    private final Map<UUID, Todo> todos = new HashMap<>();
+public interface TodoRepository {
+    Todo save(Todo todo);
 
-    public Todo save(Todo todo) {
-        todos.put(todo.getId(), todo);
-
-        return todo;
-    }
-
-    public List<Todo> findAll(){
-        return todos.values()
-                .stream()
-                .toList();
-    }
+    List<Todo> findAll();
 }
